+++ CHANGELOG.md
@@ -1,9 +1,18 @@
 # Changelog
 
+## 2.0.0
+
+### Breaking changes
+
+- Updated dependency to `@animoca/ethereum-contracts-assets@2.0.0`.
+- Removed Pausable feature from TokenLaunchpadVouchers.
+
 ## 1.1.0
 
 ### Improvements
- * Remove mintable ERC165 interface from TokenLaunchpadVouchers.
+
+- Removed mintable ERC165 interface from TokenLaunchpadVouchers.
 
 ## 1.0.0
- * Initial release.
+
+- Initial release.